ReadType.TRY_CACHE finished. Should work for singal node.

+++ src/tachyon/Constants.java
@@ -21,7 +21,7 @@ package tachyon;
 public class Constants {
   public static final int KB = 1024;
   public static final int MB = KB * 1024;
-  public static final int GB = MB * 1024;
+  public static final long GB = MB * 1024;
   public static final long TB = GB * 1024L;
   public static final long PB = TB * 1024L;
